refactor: bookmarks service export and import
