+++ src/shared/services/bookmarks-views-plays-service/index.ts
@@ -1,0 +1,3 @@
+export * from './bookmarks-views-plays-service.const';
+export * from './bookmarks-views-plays-service.gql';
+export * from './bookmarks-views-plays-service';
